refactor: restructure time tracker components and utilities
- Break down TimeEntryForm into smaller, focused sub-components
- Separate TimeRecordsList functionality into individual components
- Reorganize time-related utilities into dedicated time/ directory
- Remove unused timezone helper files
- Add custom hooks for form management and handlers
- Improve component maintainability and readability
