+++ src/features/time-tracker/utils/index.ts
@@ -4,11 +4,8 @@ export * from "./formatting";
 // 統計計算相關工具
 export * from "./statisticsCalculation";
 
-// 時間計算相關工具
-export * from "./timeCalculation";
-
-// 時區處理相關工具 (推薦使用)
-export * from "./timezoneHelpers";
+// 時間相關工具 (推薦使用)
+export * from "./time";
 
 // 驗證相關工具
 export * from "./validation";
